refactor:[BE]response 클래스 lombok추가
