+++ backend/src/main/java/com/example/webduck/webtoon/service/port/KoreaWebtoonResponse.java
@@ -3,12 +3,15 @@ package com.example.webduck.webtoon.service.port;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
 public class KoreaWebtoonResponse {
     private List<WebtoonKor> webtoons;
 
-
+    @Setter
+    @NoArgsConstructor
     @AllArgsConstructor
     @Getter
     public static class WebtoonKor {
